move test to its own file to avoid unrelated debug-ASAN timeouts
